hodl_django/requirements.txt
asgiref==3.5.1
Django==4.0.4
djangorestframework==3.13.1
Markdown==3.3.6
pytz==2022.1
sqlparse==0.4.2
